% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lists.R
\name{get_stock_list}
\alias{get_stock_list}
\title{Function calling a list of stocks for which to call data}
\usage{
get_stock_list()
}
\value{
A list with the stock identifiers for Quandl
}
\description{
Function calling a list of stocks for which to call data
}

