% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-add-ons.R
\name{go_to_the_diogenes_club}
\alias{go_to_the_diogenes_club}
\title{Launch app by going to the Diogenes Club}
\usage{
go_to_the_diogenes_club()
}
\description{
Launch app by going to the Diogenes Club
}

