% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-add-ons.R
\name{loading_indicator}
\alias{loading_indicator}
\title{Widget indicating when data in the Shiny session is being loaded, (typically
data being fetched from a webservice). Heavily based on the busy-indicator
function implemented in the ShinySky package.}
\usage{
loading_indicator(wait_text = "Data is being loaded",
  gif = "mycroftr/loading-indicator/loading-indicator.gif",
  wait_time = 1000)
}
\arguments{
\item{wait_text}{A character string with the text to be displayed while
loading.}

\item{gif}{path to the gif that is to be diaplayed.}

\item{wait_time}{The amount of time to wait before showing the loading
indicator. The default time is 500 corresponding to 0.5 second.}
}
\description{
Widget indicating when data in the Shiny session is being loaded, (typically
data being fetched from a webservice). Heavily based on the busy-indicator
function implemented in the ShinySky package.
}

