% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highstocks.R
\name{plot_candlestick}
\alias{plot_candlestick}
\title{Function generating a Highstocks candlestick chart with corresponding volume
bar plot. Essentially a wrapper of the Highcharter library.}
\usage{
plot_candlestick(in_data)
}
\arguments{
\item{in_data}{A named list element containing a data frame.}
}
\value{
A Highcharter object
}
\description{
Function generating a Highstocks candlestick chart with corresponding volume
bar plot. Essentially a wrapper of the Highcharter library.
}

