% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highstocks.R
\name{plot_line}
\alias{plot_line}
\title{Function generating a Highstocks line chart. Essentially a wrapper of the
Highcharter library.}
\usage{
plot_line(in_data)
}
\arguments{
\item{in_data}{A named list element containing a data frame.}
}
\value{
A Highcharter object
}
\description{
Function generating a Highstocks line chart. Essentially a wrapper of the
Highcharter library.
}

